Reworked polyfill support for Zend\Session
- Created Zend\Session\Container and Zend\Session\Storage\SessionArrayStorage
  namespaces. Each contains PhpLegacyCompatibility and PhpReferenceCompatibility
  classes. The classes with the same name as the new namespaces each use
  conditionals to execute class_alias() commands in order to perform
  version-specific polyfills.
- Removed compatibility/autoload.php from composer.json autoload.files; no
  longer necessary.
- All tests are running!
