+++ library/Zend/Session/Container.php
@@ -7,34 +7,8 @@
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
-namespace Zend\Session;
-
-/**
- * Session storage container
- *
- * Allows for interacting with session storage in isolated containers, which
- * may have their own expiries, or even expiries per key in the container.
- * Additionally, expiries may be absolute TTLs or measured in "hops", which
- * are based on how many times the key or container were accessed.
- */
-class Container extends AbstractContainer
-{
-    /**
-     * Retrieve a specific key in the container
-     *
-     * @param  string $key
-     * @return mixed
-     */
-    public function &offsetGet($key)
-    {
-        $ret = null;
-        if (!$this->offsetExists($key)) {
-            return $ret;
-        }
-        $storage = $this->getStorage();
-        $name    = $this->getName();
-        $ret =& $storage[$name][$key];
-
-        return $ret;
-    }
+if (version_compare(PHP_VERSION, '5.3.4', 'lt')) {
+    class_alias('Zend\Session\Container\PhpLegacyCompatibility', 'Zend\Session\Container');
+} else {
+    class_alias('Zend\Session\Container\PhpReferenceCompatibility', 'Zend\Session\Container');
 }
