+++ library/Zend/Session/Storage/SessionArrayStorage.php
@@ -7,32 +7,9 @@
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
-namespace Zend\Session\Storage;
-
-/**
- * Session storage in $_SESSION
- */
-class SessionArrayStorage extends AbstractSessionArrayStorage
-{
-    /**
-     * Get Offset
-     *
-     * @param  mixed $key
-     * @return mixed
-     */
-    public function &__get($key)
-    {
-        return $_SESSION[$key];
-    }
-
-    /**
-     * Offset Get
-     *
-     * @param  mixed $key
-     * @return mixed
-     */
-    public function &offsetGet($key)
-    {
-        return $_SESSION[$key];
-    }
+if (version_compare(PHP_VERSION, '5.3.4', 'lt')) {
+    class_alias('Zend\Session\Storage\SessionArrayStorage\PhpLegacyCompatibility', 'Zend\Session\Storage\SessionArrayStorage');
+} else {
+    class_alias('Zend\Session\Storage\SessionArrayStorage\PhpReferenceCompatibility', 'Zend\Session\Storage\SessionArrayStorage');
 }
+
